requirements.txt
telebot
openai
langchain
python-dotenv
pypdf
chromadb
tiktoken
tabulate
langid
deep-translator
SpeechRecognition==3.10.0
python-telegram-bot